chore(sdk): Add test for if and concat placeholder and validate artifact channel usage (#8120)
* test if and concat placeholder and validate artifact channel usage

* fix nit in read write test config

* resolve comments

* resolve comments and move test cases

* resolve comments and add placeholder validation

* remove unused imports

* update release file
