+++ sdk/RELEASE.md
@@ -7,6 +7,7 @@
 ### For Pipeline Authors
 
 ### For Component Authors
+* Add support for containerized component [\#8066](https://github.com/kubeflow/pipelines/pull/8066)
 
 ## Deprecations
