+++ sdk/python/kfp/compiler/test_data/components/container_with_if_placeholder.yaml
@@ -1,0 +1,55 @@
+components:
+  comp-container-with-if-placeholder:
+    executorLabel: exec-container-with-if-placeholder
+    inputDefinitions:
+      parameters:
+        optional_input:
+          defaultValue: default
+          parameterType: STRING
+    outputDefinitions:
+      artifacts:
+        dataset:
+          artifactType:
+            schemaTitle: system.Dataset
+            schemaVersion: 0.0.1
+      parameters:
+        output_path:
+          parameterType: STRING
+deploymentSpec:
+  executors:
+    exec-container-with-if-placeholder:
+      container:
+        args:
+        - --output_path
+        - '{{$.outputs.parameters[''output_path''].output_file}}'
+        command:
+        - my_program
+        - '{"IfPresent": {"InputName": "optional_input", "Then": ["{{$.inputs.parameters[''optional_input'']}}"],
+          "Else": ["bye"]}}'
+        - --dataset
+        - '{"IfPresent": {"InputName": "optional_input", "Then": ["{{$.outputs.artifacts[''dataset''].uri}}"],
+          "Else": ["bye"]}}'
+        image: python:3.7
+pipelineInfo:
+  name: container-with-if-placeholder
+root:
+  dag:
+    tasks:
+      container-with-if-placeholder:
+        cachingOptions:
+          enableCache: true
+        componentRef:
+          name: comp-container-with-if-placeholder
+        inputs:
+          parameters:
+            optional_input:
+              componentInputParameter: optional_input
+        taskInfo:
+          name: container-with-if-placeholder
+  inputDefinitions:
+    parameters:
+      optional_input:
+        defaultValue: default
+        parameterType: STRING
+schemaVersion: 2.1.0
+sdkVersion: kfp-2.0.0-beta.1
