+++ sdk/python/kfp/components/component_factory_test.py
@@ -14,9 +14,7 @@
 
 import unittest
 
-from kfp import dsl
 from kfp.components import component_factory
-from kfp.components import placeholders
 
 
 class TestGetPackagesToInstallCommand(unittest.TestCase):
@@ -46,39 +44,3 @@ class TestGetPackagesToInstallCommand(unittest.TestCase):
         concat_command = ' '.join(command)
         for package in packages_to_install + pip_index_urls:
             self.assertTrue(package in concat_command)
-
-
-class TestContainerComponentArtifactChannel(unittest.TestCase):
-
-    def test_correct_placeholder_and_attribute_error(self):
-        in_channel = component_factory.ContainerComponentArtifactChannel(
-            'input', 'my_dataset')
-        out_channel = component_factory.ContainerComponentArtifactChannel(
-            'output', 'my_result')
-        self.assertEqual(in_channel.uri,
-                         placeholders.InputUriPlaceholder('my_dataset'))
-        self.assertEqual(out_channel.path,
-                         placeholders.OutputPathPlaceholder('my_result'))
-        self.assertRaisesRegex(AttributeError,
-                               r'Cannot access artifact attribute "name"',
-                               lambda: in_channel.name)
-        self.assertRaisesRegex(AttributeError,
-                               r'Cannot access artifact attribute "channel"',
-                               lambda: out_channel.channel)
-
-
-class TestContainerComponentFactory(unittest.TestCase):
-
-    def test_raise_error_if_access_artifact_by_itself(self):
-
-        def comp_with_artifact_input(dataset: dsl.Input[dsl.Dataset]):
-            return dsl.ContainerSpec(
-                image='gcr.io/my-image',
-                command=['sh', 'run.sh'],
-                args=[dataset])
-
-        self.assertRaisesRegex(
-            TypeError,
-            r'Cannot access artifact by itself in the container definition.',
-            component_factory.create_container_component_from_func,
-            comp_with_artifact_input)
